Restricting access to CRUD view

+++ Controller/AdminController.php
@@ -1,28 +1,39 @@
 <?php
-    namespace Controller;
+    declare(strict_types=1);
 
+    use model\classes\AccessControl;
     use model\classes\Query;
     use model\classes\Validate;
 
     class AdminController
     {
-        private object $dbcon;
+        use AccessControl;        
 
-        public function __construct(object $dbcon)
+        public function __construct(private object $dbcon = DB_CON)
         {
-            $this->dbcon = $dbcon;
+                        
         }
 
         /** Show user index */
         public function index(): void
         {
-            $query = "SELECT * FROM user INNER JOIN roles ON user.id_role = roles.id_roles";
+            try {
+                /** Test access */
+                if(!$this->testAccess(['ROLE_ADMIN'])) throw new Exception("You must be admin to access.", 1);
+
+                $query = "SELECT * FROM user INNER JOIN roles ON user.id_role = roles.id_roles";
 				
-            $stm = $this->dbcon->pdo->prepare($query);                                        
-            $stm->execute();       
-            $rows = $stm->fetchAll();
-            $stm->closeCursor();
-            //$this->dbcon = null;
+                $stm = $this->dbcon->pdo->prepare($query);                                        
+                $stm->execute();       
+                $rows = $stm->fetchAll();
+                $stm->closeCursor();                
+                
+            } catch (\Throwable $th) {
+                $error_msg = "<p>Descripción del error: <span class='error'>{$th->getMessage()}</span></p>";
+					include(SITE_ROOT . "/../view/database_error.php");
+            }
+
+            
 
             include(SITE_ROOT . "/../view/admin/index_view.php");
         }
@@ -30,59 +41,63 @@
         /** Create new user */
         public function new(): void
         {
-            $user_name = $_REQUEST['user_name'] ?? "";
-				$password = $_REQUEST['password'] ?? "";
-				$email = $_REQUEST['email'] ?? "";
-	
-				try {
-					if (!empty($user_name) && !empty($password) && !empty($email)) {
-						$query = new Query();
-	
-						$rows = $query->selectAllBy("user", "email", $email, $this->dbcon);
-	
-						if ($rows) {
-							$error_msg = "<p class='error'>El email '{$email}' ya está registrado</p>";
-							include(SITE_ROOT . "/../view/admin/user_new_view.php");											
-						}
-						else {
-							$query = "INSERT INTO user (user_name, password, email) VALUES (:name, :password, :email)";                 
-		
-							$stm = $this->dbcon->pdo->prepare($query); 
-							$stm->bindValue(":name", $user_name);
-							$stm->bindValue(":password", password_hash($password, PASSWORD_DEFAULT));
-							$stm->bindValue(":email", $email);              
-							$stm->execute();       				
-							$stm->closeCursor();
-							$this->dbcon = null;
-			
-							$success_msg = "<p>El usuario se ha registrado correctamente</p>";
-							include(SITE_ROOT . "/../view/database_error.php");
-						}										
-					}
-					else {
-						include(SITE_ROOT . "/../view/admin/user_new_view.php");
-					}
-				} catch (\Throwable $th) {			
-					$error_msg = "<p>Hay problemas al conectar con la base de datos, revise la configuración 
-							de acceso.</p><p>Descripción del error: <span class='error'>{$th->getMessage()}</span></p>";
-					include(SITE_ROOT . "/../view/database_error.php");				
-				}			
+            $user_name  = $_REQUEST['user_name'] ?? "";
+            $password   = $_REQUEST['password'] ?? "";
+            $email      = $_REQUEST['email'] ?? "";
+
+            try {
+                /** Test access */
+                if(!$this->testAccess(['ROLE_ADMIN'])) throw new Exception("You must be admin to access.", 1);
+
+                if (!empty($user_name) && !empty($password) && !empty($email)) {
+                    $query = new Query();
+
+                    $rows = $query->selectAllBy("user", "email", $email, $this->dbcon);
+
+                    if ($rows) {
+                        $error_msg = "<p class='error'>El email '{$email}' ya está registrado</p>";
+                        include(SITE_ROOT . "/../view/admin/user_new_view.php");											
+                    }
+                    else {
+                        $query = "INSERT INTO user (user_name, password, email) VALUES (:name, :password, :email)";                 
+    
+                        $stm = $this->dbcon->pdo->prepare($query); 
+                        $stm->bindValue(":name", $user_name);
+                        $stm->bindValue(":password", password_hash($password, PASSWORD_DEFAULT));
+                        $stm->bindValue(":email", $email);              
+                        $stm->execute();       				
+                        $stm->closeCursor();
+                        $this->dbcon = null;
+        
+                        $success_msg = "<p>El usuario se ha registrado correctamente</p>";
+                        include(SITE_ROOT . "/../view/database_error.php");
+                    }										
+                }
+                else {
+                    include(SITE_ROOT . "/../view/admin/user_new_view.php");
+                }
+            } catch (\Throwable $th) {			
+                $error_msg = "<p>Descripción del error: <span class='error'>{$th->getMessage()}</span></p>";
+                include(SITE_ROOT . "/../view/database_error.php");				
+            }			
         }
 
         public function show(): void
         {
-            $id_user = $_REQUEST['id_user'];
+            $id_user = $_REQUEST['id_user'] ?? "";
 	
             $query = new Query();
 
             try {
+                /** Test access */
+                if(!$this->testAccess(['ROLE_ADMIN'])) throw new Exception("You must be admin to access.", 1);
+
                 $user = $query->selectOneBy("user", "id_user", $id_user, $this->dbcon);
 
                 include(SITE_ROOT . "/../view/admin/user_show_view.php");
                 
             } catch (\Throwable $th) {
-                $error_msg = "<p>Hay problemas al conectar con la base de datos, revise la configuración 
-                    de acceso.</p><p>Descripción del error: <span class='error'>{$th->getMessage()}</span></p>";
+                $error_msg = "<p>Descripción del error: <span class='error'>{$th->getMessage()}</span></p>";
                 include(SITE_ROOT . "/../view/database_error.php");					
             }	
         }
@@ -95,6 +110,9 @@
             $email = $_REQUEST['email'] ?? "";
 
             try {
+                /** Test access */
+                if(!$this->testAccess(['ROLE_ADMIN'])) throw new Exception("You must be admin to access.", 1);
+
                 $query = new Query();
                 $query->updateRegistry("user", $user_name, $email, $id_user, $this->dbcon);
 
@@ -103,8 +121,7 @@
                 include(SITE_ROOT . "/../view/database_error.php");
 
             } catch (\Throwable $th) {			
-                $error_msg = "<p>Hay problemas al conectar con la base de datos, revise la configuración 
-                        de acceso.</p><p>Descripción del error: <span class='error'>{$th->getMessage()}</span></p>";
+                $error_msg = "<p>Descripción del error: <span class='error'>{$th->getMessage()}</span></p>";
                 include(SITE_ROOT . "/../view/database_error.php");				
             }
         }
@@ -119,6 +136,9 @@
             $newPassword = $validate->test_input($_REQUEST['new_password'] ?? "");
 
             try {
+                /** Test access */
+                if(!$this->testAccess(['ROLE_ADMIN'])) throw new Exception("You must be admin to access.", 1);
+
                 if (!empty($password) && !empty($id_user) && !empty($newPassword)) {
                     if ($password !== $newPassword) {
                         $error_msg = "<p class='alert alert-danger text-center'>Las contraseñas no son iguales</p>";
@@ -131,8 +151,7 @@
                     
                 }
             } catch (\Throwable $th) {
-                $error_msg = "<p>Hay problemas al conectar con la base de datos, revise la configuración 
-                        de acceso.</p><p>Descripción del error: <span class='error'>{$th->getMessage()}</span></p>";
+                $error_msg = "<p>Descripción del error: <span class='error'>{$th->getMessage()}</span></p>";
                 include(SITE_ROOT . "/../view/database_error.php");
             }
 
@@ -142,9 +161,12 @@
         /** Deleting a user from the database. */
         public function delete(): void
         {
-            $id_user = $_REQUEST['id_user'];
+            $id_user = $_REQUEST['id_user'] ?? "";
 	
             try {
+                /** Test access */
+                if(!$this->testAccess(['ROLE_ADMIN'])) throw new Exception("You must be admin to access.", 1);
+
                 $query = new Query();
                 $query->deleteRegistry("user", $id_user, $this->dbcon);
 
@@ -153,8 +175,7 @@
                 include(SITE_ROOT . "/../view/database_error.php");
 
             } catch (\Throwable $th) {
-                $error_msg = "<p>Hay problemas al conectar con la base de datos, revise la configuración 
-                        de acceso.</p><p>Descripción del error: <span class='error'>{$th->getMessage()}</span></p>";
+                $error_msg = "<p>Descripción del error: <span class='error'>{$th->getMessage()}</span></p>";
                 include(SITE_ROOT . "/../view/database_error.php");
             }
         }
